refactored shopping commands fixed Grocy method access in GrocyCache refactored keyboard response handling into custom classes
